That was pure hell. I believe I've successfully written a logIn api for you to use. you can reach it at caredb.herokuapp.com/login.php it takes three arguments: UName Pass SessionKey
If you provide a SessionKey, it will check it for validity. A valid key
will return the UName, an invalid key will return null.
If you provide a SessionKey, that key will be checked before UName and Pass.

If you provide a UName and Pass and NO session key, it will check it for
validity. A valid combination will create a new session key and return it
as a json. An invalid key will return null.

All other kinds of entries will return null.

I have refreshed the /Reset.php file and it should work with the current
database prototype. Using it will drop all tables and entries.

I will add a view /SessionKeys.php that will return a json of the
SessionKeys table, for debugging.

Good night