Just for David! I noticed you had a function for addPillRecord() I tried to copy it's functionality for the online database just go to caredb.herokuapp.com/addPillLog.php?alert_time=someTime&log_time=someTime&message=someMessage&action_taken=someString&SessionKey=someValidSessionKey
I was able to successfully insert a record by using it, so it is
functional.

you'll have to have a valid session key though, which can be achieved by
visiting
caredb.com/?UName=dscianni&Pass=david
just copy and paste that for testing, and it will work

You only get 1 SessionKey at a time though, so if you log in again you
have to get the new SessionKey.

I have yet to do anything that will allow you to retrieve the pillRecords
however.

Additional updates:
php queries not related to app functionality (ie: Users.php that just
shows you what's in the users table) have been moved to /devutils. So for
example you can go to caredb.herokuapp.com/devutils/Users.php to get a
json that describes the entries in the users table.

Changed the way login works. Now it's going to be a function accessible
from php only. The webservice will handle the response if you aren't
logged in.
loginCopy.php is available with the old functionality if you're testing
it. You can also use that page to get a SessionKey but it will probably be
dissappearing shortly.

I did manage to get my offline build environment tip-top, so no problems
there. This is much nicer because instead of waiting like a minute to fix
a missing semicolon and see the update, I can see it as soon as I press
refresh... So Yay... Win...

Now it's time for sleep.
Oh, PS. For a limited time there's some coments in the index.php that are
similar to this commit, but may be more detailed.

Those will dissappear soon as well but maybe you'll find them useful for
now.